pyproject.toml
[tool.poetry]
name = "ecom-test"
version = "0.1.0"
description = ""
authors = ["palas1k <liserg008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
pymongo = "^4.10.1"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
